Remove old JS, setup webpack production command
