+++ js/components/StatsContainer.js
