+++ js/components/VerticalLabel.js
