--- js/react/index.js
@@ -1,18 +1,0 @@
-import ReactDOM from 'react-dom'
-import React from 'react'
-import App from './App.js'
-import { createStore, applyMiddleware } from 'redux'
-import { Provider } from 'react-redux'
-import reducers from './reducers'
-import reduxThunk from 'redux-thunk'
-import 'react-select/dist/react-select.css'
-
-const createStoreWithMiddleware = applyMiddleware(reduxThunk)(createStore)
-const store = createStoreWithMiddleware(reducers)
-
-ReactDOM.render(
-  <Provider store={store}>
-    <App />
-  </Provider>,
-  document.getElementById('react-root')
-)
