+++ js/reducers/gdqReducers.js
