+++ react.html
@@ -9,7 +9,6 @@
 
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css">
   {% include favicon.html %}
-  {% include libraries.html %}
   {% include google_analytics.html %}
   {% include header.html %}
