+++ webpack.config.js
@@ -4,7 +4,6 @@ var APP_DIR = path.resolve(__dirname, 'js');
 module.exports = {
   entry: {
     index: APP_DIR + "/index.js",
-    index_react: APP_DIR + "/react/index.js"
   },
   output: {
     filename: "./dist/[name].bundle.js",
